[C-1551] Add collection screen skeleton
