+++ packages/mobile/src/components/navigation-container/NavigationContainer.tsx
@@ -141,6 +141,7 @@ const NavigationContainer = (props: NavigationContainerProps) => {
       }
     },
     getStateFromPath: (path, options) => {
+      console.log('hmmm', path)
       // Add leading slash if it is missing
       if (path[0] !== '/') path = `/${path}`
 
@@ -191,13 +192,19 @@ const NavigationContainer = (props: NavigationContainerProps) => {
           }
         }
 
-        // If the path is a playlist or album
-        if (path.match(/^\/[^/]+\/(playlist|album)\/[^/]+$/)) {
+        if (path.match(/^\/[^/]+\/playlist\/[^/]+$/)) {
           // set the path as `collection`
+          console.log('playlist')
           path = path.replace(
-            /(^\/[^/]+\/)(playlist|album)(\/[^/]+$)/,
+            /(^\/[^/]+\/)(playlist)(\/[^/]+$)/,
             '$1collection$3'
           )
+          path = `${path}?collectionType=playlist`
+          console.log('path!', path)
+        } else if (path.match(/^\/[^/]+\/album\/[^/]+$/)) {
+          // set the path as `collection`
+          path = path.replace(/(^\/[^/]+\/)(album)(\/[^/]+$)/, '$1collection$3')
+          path = `${path}?collectionType=album`
         }
       }
