+++ packages/mobile/src/components/skeletons/index.ts
@@ -1,0 +1,1 @@
+export * from './EntitySkeleton'
