+++ packages/mobile/src/screens/collection-screen/CollectionScreen.tsx
@@ -35,6 +35,7 @@ import type { SearchPlaylist, SearchUser } from 'app/store/search/types'
 import { makeStyles } from 'app/styles'
 
 import { CollectionScreenDetailsTile } from './CollectionScreenDetailsTile'
+import { CollectionScreenSkeleton } from './CollectionScreenSkeleton'
 
 const { setFavorite } = favoritesUserListActions
 const { setRepost } = repostsUserListActions
@@ -64,7 +65,12 @@ export const CollectionScreen = () => {
   const dispatch = useDispatch()
 
   // params is incorrectly typed and can sometimes be undefined
-  const { id: idParam, searchCollection, collectionName } = params ?? {}
+  const {
+    id: idParam,
+    searchCollection,
+    collectionName,
+    collectionType
+  } = params ?? {}
 
   const id = useMemo(() => {
     if (collectionName) {
@@ -96,25 +102,22 @@ export const CollectionScreen = () => {
   const collection = cachedCollection ?? searchCollection
   const user = cachedUser ?? searchCollection?.user
 
-  if (!collection || !user) {
-    // TODO: add collection-screen skeleton
-    return null
-  }
-
-  return <CollectionScreenComponent collection={collection} user={user} />
+  return !collection || !user ? (
+    <CollectionScreenSkeleton collectionType={collectionType} />
+  ) : (
+    <CollectionScreenComponent collection={collection} user={user} />
+  )
 }
 
 type CollectionScreenComponentProps = {
   collection: Collection | SearchPlaylist
   user: User | SearchUser
 }
-const CollectionScreenComponent = ({
-  collection,
-  user
-}: CollectionScreenComponentProps) => {
+const CollectionScreenComponent = (props: CollectionScreenComponentProps) => {
   const styles = useStyles()
   const dispatch = useDispatch()
   const navigation = useNavigation()
+  const { collection, user } = props
   const {
     _is_publishing,
     description,
