+++ packages/mobile/src/screens/collection-screen/CollectionScreenSkeleton.tsx
@@ -1,0 +1,26 @@
+import { View } from 'react-native'
+
+import { StaticSkeleton } from 'app/components/skeleton/Skeleton'
+import { EntitySkeleton } from 'app/components/skeletons'
+import { TrackListItemSkeleton } from 'app/components/track-list/TrackListItemSkeleton'
+
+type CollectionScreenSkeletonProps = {
+  collectionType?: 'playlist' | 'album'
+}
+
+export const CollectionScreenSkeleton = (
+  props: CollectionScreenSkeletonProps
+) => {
+  const { collectionType } = props
+
+  return (
+    <EntitySkeleton entityType={collectionType}>
+      <View>
+        <StaticSkeleton />
+        <StaticSkeleton />
+      </View>
+      <TrackListItemSkeleton />
+      <TrackListItemSkeleton />
+    </EntitySkeleton>
+  )
+}
